Updated how processMemoryOption() reads user input

+++ MemoryAllocation/MemoryBlock.java
@@ -1,10 +1,11 @@
-public class MemoryBlock {
-    private int startAddress;
-    private int size;
+public class MemoryBlock 
+{
+    private long startAddress;
+    private long size;
     private boolean isFree;
     private String processId;
 
-    public MemoryBlock(int startAddress, int size, boolean isFree, String processId) 
+    public MemoryBlock(long startAddress, long size, boolean isFree, String processId) 
     {
         this.startAddress = startAddress;
         this.size = size;
@@ -12,22 +13,22 @@ public class MemoryBlock {
         this.processId = processId;
     }
 
-    public int getStartAddress() 
+    public long getStartAddress() 
     {
         return startAddress;
     }
 
-    public void setStartAddress(int startAddress) 
+    public void setStartAddress(long startAddress) 
     {
         this.startAddress = startAddress;
     }
 
-    public int getSize() 
+    public long getSize() 
     {
         return size;
     }
 
-    public void setSize(int size) 
+    public void setSize(long size) 
     {
         this.size = size;
     }
